refactor: shared isInClusterModeSelector
