+++ src/components/molecules/CheckedResourcesActionsMenu/CheckedResourcesActionsMenu.tsx
@@ -6,8 +6,7 @@ import {CloseOutlined, ExclamationCircleOutlined} from '@ant-design/icons';
 
 import {makeApplyMultipleResourcesText} from '@constants/makeApplyText';
 
-import {isInClusterModeSelector, kubeConfigContextColorSelector} from '@redux/appConfig';
-import {selectKubeContext} from '@redux/cluster/selectors';
+import {kubeConfigContextColorSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setAlert} from '@redux/reducers/alert';
 import {editorHasReloadedSelectedPath, uncheckAllResourceIds} from '@redux/reducers/main';
@@ -20,8 +19,9 @@ import {AlertEnum} from '@shared/models/alert';
 import {AppDispatch} from '@shared/models/appDispatch';
 import {ResourceIdentifier, ResourceMeta} from '@shared/models/k8sResource';
 import {Colors} from '@shared/styles/colors';
+import {selectKubeContext} from '@shared/utils/cluster/selectors';
 import {isDefined} from '@shared/utils/filter';
-import {isInPreviewModeSelector} from '@shared/utils/selectors';
+import {isInClusterModeSelector, isInPreviewModeSelector} from '@shared/utils/selectors';
 
 import ModalConfirmWithNamespaceSelect from '../ModalConfirmWithNamespaceSelect';
 import * as S from './CheckedResourcesActionMenu.styled';
