+++ src/components/organisms/ActionsPane/Scale/Scale.tsx
@@ -5,7 +5,7 @@ import {Col, InputNumber, Modal, Row, Tooltip} from 'antd';
 import {TOOLTIP_DELAY} from '@constants/constants';
 import {ScaleTooltip} from '@constants/tooltips';
 
-import {isInClusterModeSelector, kubeConfigContextSelector, kubeConfigPathSelector} from '@redux/appConfig';
+import {kubeConfigContextSelector, kubeConfigPathSelector} from '@redux/appConfig';
 import {connectCluster} from '@redux/cluster/thunks/connect';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {closeScaleModal, openScaleModal} from '@redux/reducers/ui';
@@ -13,6 +13,8 @@ import {useSelectedResource} from '@redux/selectors/resourceSelectors';
 import scaleDeployment from '@redux/services/scaleDeployment';
 
 import {PrimaryButton} from '@atoms';
+
+import {isInClusterModeSelector} from '@shared/utils/selectors';
 
 const Scale: React.FC = () => {
   const dispatch = useAppDispatch();
