+++ src/components/organisms/GlobalModals/GlobalModals.tsx
@@ -5,7 +5,7 @@ import {Image, Modal} from 'antd';
 import fs from 'fs';
 import semver from 'semver';
 
-import {isInClusterModeSelector, setDeleteProject} from '@redux/appConfig';
+import {setDeleteProject} from '@redux/appConfig';
 import {toggleForm} from '@redux/forms';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setAlert} from '@redux/reducers/alert';
@@ -17,7 +17,8 @@ import PerformanceIcon from '@assets/PerformanceIcon.svg';
 
 import {AlertEnum} from '@shared/models/alert';
 import {Project} from '@shared/models/config';
-import {electronStore} from '@shared/utils';
+import electronStore from '@shared/utils/electronStore';
+import {isInClusterModeSelector} from '@shared/utils/selectors';
 
 const GitCloneModal = React.lazy(() => import('@organisms/GitCloneModal'));
