+++ src/components/organisms/PaneManager/PaneManager.tsx
@@ -2,7 +2,7 @@ import React, {useCallback, useMemo} from 'react';
 
 import {Skeleton} from 'antd';
 
-import {activeProjectSelector, isInClusterModeSelector} from '@redux/appConfig';
+import {activeProjectSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setPaneConfiguration, toggleLeftMenu} from '@redux/reducers/ui';
 
@@ -12,6 +12,7 @@ import {EmptyDashboard} from '@organisms/Dashboard/EmptyDashboard';
 import {useMainPaneDimensions} from '@utils/hooks';
 
 import {ResizableColumnsPanel, ResizableRowsPanel} from '@monokle/components';
+import {isInClusterModeSelector} from '@shared/utils/selectors';
 
 import ProblemPane from '../ProblemPane';
 import StartPage from '../StartPage';
