+++ src/redux/cluster/selectors.ts
@@ -5,37 +5,16 @@ import {TypedUseSelectorHook} from 'react-redux';
 import {uniq} from 'lodash';
 import {createSelector} from 'reselect';
 
-import {kubeConfigPathSelector} from '@redux/appConfig';
 import {useAppSelector} from '@redux/hooks';
 
 import type {ClusterState} from '@shared/models/clusterState';
 import {ModernKubeConfig} from '@shared/models/config';
 import type {RootState} from '@shared/models/rootState';
+import {selectKubeconfig} from '@shared/utils/cluster/selectors';
 import {isDefined} from '@shared/utils/filter';
 
 export const useClusterSelector: TypedUseSelectorHook<ClusterState> = selector =>
   useAppSelector(state => selector(state.cluster));
-
-export const selectKubeconfig = (
-  state: RootState,
-  options: {kubeconfig?: string} = {}
-): ModernKubeConfig | undefined => {
-  const kubeconfigPath = options.kubeconfig ?? kubeConfigPathSelector(state);
-  if (!kubeconfigPath) return undefined;
-  const kubeconfig = state.cluster.kubeconfigs[kubeconfigPath];
-  return kubeconfig;
-};
-
-export const selectKubeContext = (
-  state: RootState,
-  options: {kubeconfig?: string; context?: string} = {}
-): Context | undefined => {
-  const kubeconfig = selectKubeconfig(state, options);
-  if (!kubeconfig?.isValid) return undefined;
-  const contextName = options.context ?? kubeconfig.currentContext;
-  const context = kubeconfig.contexts.find(c => c.name === contextName);
-  return context;
-};
 
 export function getContext(kubeconfig: ModernKubeConfig | undefined, context?: string): Context | undefined {
   if (!kubeconfig?.isValid) return undefined;
